+++ docs/CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## [Unreleased]
 ### Fixed
-- Added the pairing\_file\_name paramter to read\_txt parser
 
+## [0.0.9] - 2023-01-10
+- Added the pairing\_file\_name paramter to read\_txt parser
+- Added a FAQ to the readthedocs
+- Added protein group level FDR
 
 ## [0.0.6] - 2023-06-02
 ### Added